Move some files to `/lib` Better organization
